fix: make build scripts not rely on git repo submodules
Doing so is troublesome and does not work when building dependencies
from crates.io, as any git repository information is lost in the
process. However, `cargo publish` is smart enough to follow symlinks to
our vendor submodules, and treat them as if they were package files, so
let's use that to our advantage.

As a side bonus, the git2 build-time dependency is dropped.