docs(encoder_impl): explicitly document audio block format at encode_audio_block
I think the audio block format was pretty clear already given the
documentation of VorbisEncoder, VorbisDecoder and the examples, but when
reading rustdoc between the lines that might not be so obvious at a
first glance. Let's explicitly mention the audio block format here to
prevent reasonable people from guessing and wasting time.