docs(vorbis_rs): clarify that Vorbis{Decoder,Encoder}::new do I/O
From the name of the methods alone, it may not be so obvious at a first
glance that these methods do I/O operations to read Ogg Vorbis metadata
or write Ogg Vorbis headers. A reasonable developer may think that these
are delayed to the first decode or encode operation, for example.

Let's clarify their side effects better in the documentation.