feat: add `VorbisEncoderBuilder`, optional stream serial generation
As suggested in #9, instantiating a `VorbisEncoder` can lead to somewhat
ugly code in the client side, as a `VorbisEncoder` can be configured
with a fair amount of options. However, most of those options can be set
to default values for common cases.

To improve the ergonomics of `VorbisEncoder` instantiation, let's
delegate it to a new `VorbisEncoderBuilder`, which may optionally have
the new responsibility of generating random Ogg stream serials. In
addition, a `Default` implementation for
`VorbisBitrateManagementStrategy` was added.