fix: disable bindgen layout tests
They are mainly useful to catch unexpected layout discrepancies across
platforms, but we were already expecting such discrepancies on Windows,
due to the bindings being pregenerated on Linux.

It'd be great to still have these tests on Linux, but with current
versions of bindgen that's not doable without resorting to increasing
tech debt. (See
https://github.com/rust-lang/rust-bindgen/pull/2787#issuecomment-2506635895)